refactor: extract in call page to a separate file
